fixed structure and spacing
