+++ components/landing/hero/hero.module.scss
@@ -1,59 +1,86 @@
 @import '../../../styles/globals.scss';
 
 .container {
-	color: $textBlack;
+  color: $textBlack;
 	display: flex;
-	flex-direction: column;
-	align-items: flex-start;
-	padding: $space-lg $space-xl;
-	max-width: 800px;
-	margin: $space-md auto;
+	justify-content: space-between;
+	.aboutText {
+		width: 50%;
+		.heading {
+			display: flex;
+			.logoContainer {
+				margin: 0px 0px $space-md $space-xl;
+				background-image: linear-gradient(
+					to right bottom,
+					#f6931d,
+					#f78721,
+					#f77b26,
+					#f66f2b,
+					#f46231,
+					#f35c33,
+					#f15736,
+					#f05138,
+					#f05138,
+					#f05138,
+					#f05138,
+					#f05138
+				);
+				-webkit-border-radius: 100px / 70px;
+				border-radius: 50px / 50px;
+				border-top-left-radius: 0px;
+				border-top-right-radius: 0px;
+				border-bottom-left-radius: 150px;
+				border-bottom-right-radius: 150px;
+				width: 120px;
+				height: 130px;
+				display: flex;
+				align-items: center;
+				justify-content: center;
+			}
 
-	h2 {
-		align-self: center;
-		font-family: $primaryHeaderFont;
-	}
-
-	.about {
-		font-size: 1.2rem;
-		text-align: justify;
-		width: 600px;
-		align-self: center;
-
-		strong {
-			font-weight: bold;
+			.logo {
+				height: 75px;
+				width: 75px;
+			}
+			h3 {
+				font-family: $primaryHeaderFont;
+        font-weight: bold;
+        margin: $space-md 0px 0px $space-md;
+			}
 		}
-	}
+    .about {
+      padding: $space-lg;
+      display:flex;
+      flex-direction: column;
 
-	.image {
+      h2 {
+        color: $udSecondaryOrange;
+        font-family: $primaryHeaderFont;
+        font-weight: bold;
+        text-transform: uppercase;
+      }
+
+      p {
+        font-size: $f5;
+        font-family: $primaryTextFont;
+        line-height: 1.5;
+        padding: $space-md 0;
+        strong {
+          font-weight: bold;
+        }
+      }
+    }
+	}
+	.aboutImage {
+		background-image: linear-gradient(rgba(0, 0, 0, 0.2), rgba(0, 0, 0, 0.7)), url('/images/collab.jpg');
+		height: 400px;
 		width: 600px;
-		align-self: center;
-		margin: $space-lg;
-	}
-	p {
-		font-family: $primaryTextFont;
-		color: $subTextBlack;
-		line-height: 1.5;
-		text-align: justify;
-		padding: $space-sm 0;
-	}
+		display: flex;
+		justify-content: center;
+		align-items: center;
 
-	h3 {
-		align-self: flex-start;
-		font-family: $primaryHeaderFont;
-		font-weight: bold;
-		padding-top: $space-sm;
-	}
-
-	a {
-		color: $involvedBlue;
-		font-family: $linkFont;
-		align-self: flex-end;
-		margin-bottom: $space-xxs;
-	}
-
-	ul {
-		font-family: $secondaryHeaderFont;
-		align-self: center;
+		.imageLogo {
+			margin-top: -100px;
+		}
 	}
 }
