+++ components/landing/hero/hero.tsx
@@ -1,74 +1,51 @@
-import React, { ReactElement } from 'react';
+import React from 'react';
 import styles from './hero.module.scss';
 
-interface Props { }
+interface Props {
 
-function Hero({ }: Props): ReactElement {
-	return (
-		<div className={styles.container}>
-			<h2 className={styles.story}>Our Story</h2>
-			<p className={styles.about}>
-				We are a group of software engineers helping aspiring developers who are either formerly incarcerated or from an economically disadvantaged background. <br /> We are creatting opportunities in tech for people who might not otherwise get an opportunity <br /> We were founded in September of 2020 <br /> We
-				currently depend on <strong>100% volunteer support</strong>.
-			</p>
-			<img
-				className={styles.image}
-				src="/images/collab.jpg"
-				alt="devs collaborating"
-			/>
-			<h3>Recidivism</h3>
-			<p>
-				According to data from the U.S. Bureau of Justice Statistics (BJS), 67.8% of released state prisoners
-				were arrested for a new crime within three years, and 76.7% were arrested within five years (Durose,
-				Cooper, & Snyder, 2014). Reducing recidivism not only protects society at large, but also improves the
-				life quality of individual ex-prisoners.
-			</p>
-			<p>
-				Incarceration rates in the U.S. began increasing dramatically in the 1990s. The U.S. has the highest
-				prison population of any country, comprising 25% of the world’s prisoners.{' '}
-			</p>
-			<p>Out of a sample of 401,288 prisoners state prisoners released in 2005:</p>
-			<ul>
-				<li>68% of released prisoners were arrested within 3 years.</li>
-				<li>79% of released prisoners were arrested within 6 years.</li>
-				<li>83% of released prisoners were arrested within 9 years.</li>
-			</ul>
-			<p>Using this study as a basepoint, one could state that the national recidivism rate is 83%.</p>
-			<a href="https://www.prisonpolicy.org/research/recidivism_and_reentry/" rel="noreferrer" target="_blank">
-				recidivism and reentry on prisonpolicy.org
-			</a>
-			<a
-				href="https://prisoninsight.com/recidivism-the-ultimate-guide/#tab-con-9"
-				rel="noreferrer"
-				target="_blank"
-			>
-				recidivism the ultimate guide on prisoninsight.com
-			</a>
-			<h3>Unemployment and Poverty</h3>
-			<p>
-				Employment has long been recognized as having a negative correlation with crime (Uggen, 1999; Uggen et
-				al., 2005).
-			</p>
-			<a
-				href="https://www.ebpsociety.org/blog/education/297-employment-recidivism"
-				rel="noreferrer"
-				target="_blank"
-			>
-				employment recidivism on ebpsociety.org
-			</a>{' '}
-			<p>
-				Researchers also have found that employment is associated with reduced recidivism(Skardhamar & Telle,
-				2012). The strongest predictor for recidivism: poverty + unemployment.
-			</p>
-			<a
-				href="https://www.ojp.gov/ncjrs/virtual-library/abstracts/poverty-state-capital-and-recidivism-among-women-offenders"
-				rel="noreferrer"
-				target="_blank"
-			>
-				poverty state capital and recidivism among women offenders on ojp.gov
-			</a>{' '}
-		</div>
-	);
 }
+
+const Hero: React.FC<Props> = ({ }) => {
+	return (
+		<section className={styles.container}>
+			<div className={styles.aboutText}>
+				<div className={styles.heading}>
+					<div className={styles.logoContainer}>
+						<img
+							className={styles.logo}
+							src="/images/icon-03.png"
+							alt="Your Name"
+						/>
+					</div>
+					<h3>About</h3>
+				</div>
+				<div className={styles.about}>
+					<h2>Underdog <br />Devs</h2>
+					<p className={styles.about}>
+						We are a group of software engineers helping aspiring developers who are either formerly incarcerated or from an economically disadvantaged background. <br /> We are creatting opportunities in tech for people who might not otherwise get an opportunity <br /> We were founded in September of 2020 <br /> We
+						currently depend on 100% volunteer support
+					</p>
+				</div>
+			</div>
+			<div
+				className={styles.aboutImage}
+				style={{
+					// backgroundImage: `url(${'/images/collab.jpg'})`,
+					backgroundRepeat: 'no-repeat',
+					backgroundSize: 'cover',
+				}}
+			>
+				<img
+					className={styles.imageLogo}
+					src="/images/underdogdevs-03.png"
+					alt="Your Name"
+					height={300}
+					width={300}
+				/>
+			</div>
+
+		</section>
+	);
+};
 
 export default Hero;
