--- components/landing/hero/newHero/index.ts
@@ -1,3 +1,0 @@
-import NewHero from './NewHero';
-
-export default NewHero;
