--- components/landing/hero/newHero/newHero.module.scss
@@ -1,86 +1,0 @@
-@import '../../../../styles/globals.scss';
-
-.container {
-  color: $textBlack;
-	display: flex;
-	justify-content: space-between;
-	.aboutText {
-		width: 50%;
-		.heading {
-			display: flex;
-			.logoContainer {
-				margin: 0px 0px 10px 50px;
-				background-image: linear-gradient(
-					to right bottom,
-					#f6931d,
-					#f78721,
-					#f77b26,
-					#f66f2b,
-					#f46231,
-					#f35c33,
-					#f15736,
-					#f05138,
-					#f05138,
-					#f05138,
-					#f05138,
-					#f05138
-				);
-				-webkit-border-radius: 100px / 70px;
-				border-radius: 50px / 50px;
-				border-top-left-radius: 0px;
-				border-top-right-radius: 0px;
-				border-bottom-left-radius: 150px;
-				border-bottom-right-radius: 150px;
-				width: 120px;
-				height: 130px;
-				display: flex;
-				align-items: center;
-				justify-content: center;
-			}
-
-			.logo {
-				height: 75px;
-				width: 75px;
-			}
-			h3 {
-				font-family: $primaryHeaderFont;
-        font-weight: bold;
-        margin: 40px 0px 0px 40px;
-			}
-		}
-    .about {
-      padding: 50px;
-      display:flex;
-      flex-direction: column;
-
-      h2 {
-        color: $udSecondaryOrange;
-        font-family: $primaryHeaderFont;
-        font-weight: bold;
-        text-transform: uppercase;
-      }
-
-      p {
-        font-size: $f5;
-        font-family: $primaryTextFont;
-        line-height: 1.5;
-        padding: 50px 0;
-        strong {
-          font-weight: bold;
-        }
-      }
-    }
-	}
-	.aboutImage {
-		background-image: linear-gradient(rgba(0, 0, 0, 0.2), rgba(0, 0, 0, 0.7)), url('/images/collab.jpg');
-		height: 400px;
-		width: 600px;
-		display: flex;
-		justify-content: center;
-		align-items: center;
-
-		.imageLogo {
-			margin-top: -100px;
-		}
-	}
-}
