+++ components/landing/recidivism/index.ts
@@ -1,0 +1,3 @@
+import Recidivism from './Recidivism';
+
+export default Recidivism;
