+++ components/landing/socialMedia/socialMedia.module.scss
@@ -21,12 +21,12 @@
       display: flex;
       flex-direction: row;
       justify-content: space-between;
-      color: $involvedBlue;
+      color: $textBlack;
       font-size: 1.5rem;
   
       .icon {
-        background-color: $lightenedBlue;
-        color: $involvedBlue;
+        background-color: $udSecondaryOrange;
+        color: $udDarkGrey;
         padding: $space-xs $space-sm $space-xxxs;
         border-radius: 25px;
         margin: $space-lg $space-xs;
