+++ pages/parts/footer/Footer.tsx
@@ -12,9 +12,9 @@ function Footer({}: Props): ReactElement {
 				<div className={styles.social}>
 					<div className={styles.logo}>
 						<img
-							src="/images/ud-logo.png" // Route of the image file
-							height={75} // Desired size with correct aspect ratio
-							width={75} // Desired size with correct aspect ratio
+							src="/images/underdogdevs-01.png" // Route of the image file
+							height={150} // Desired size with correct aspect ratio
+							width={150} // Desired size with correct aspect ratio
 							alt="Your Name"
 						/>
 					</div>
